move util function from docker to util
