+++ pkg/docker/sandbox_util.go
@@ -10,22 +10,18 @@ import (
 
 	"github.com/flyteorg/flytectl/pkg/util"
 
-	"github.com/docker/docker/client"
-
-	"github.com/enescakir/emoji"
-
-	cmdUtil "github.com/flyteorg/flytectl/pkg/commandutils"
-
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/mount"
+	"github.com/docker/docker/client"
 	"github.com/docker/go-connections/nat"
+	"github.com/enescakir/emoji"
+	cmdUtil "github.com/flyteorg/flytectl/pkg/commandutils"
 	f "github.com/flyteorg/flytectl/pkg/filesystemutils"
 )
 
 var (
 	Kubeconfig              = f.FilePathJoin(f.UserHomeDir(), ".flyte", "k3s", "k3s.yaml")
-	FlytectlConfig          = f.FilePathJoin(f.UserHomeDir(), ".flyte", "config-sandbox.yaml")
 	SuccessMessage          = "Flyte is ready! Flyte UI is available at http://localhost:30081/console"
 	ImageName               = "cr.flyte.org/flyteorg/flyte-sandbox:dind"
 	FlyteSandboxClusterName = "flyte-sandbox"
@@ -41,32 +37,6 @@ var (
 		},
 	}
 )
-
-// SetupFlyteDir will create .flyte dir if not exist
-func SetupFlyteDir() error {
-	if err := os.MkdirAll(f.FilePathJoin(f.UserHomeDir(), ".flyte"), 0755); err != nil {
-		return err
-	}
-	return nil
-}
-
-// SetupConfig download the flyte sandbox config
-func SetupConfig() error {
-	return util.WriteIntoFile([]byte(util.ConfigTemplate), FlytectlConfig)
-}
-
-// ConfigCleanup will remove the sandbox config from flyte dir
-func ConfigCleanup() error {
-	err := os.Remove(FlytectlConfig)
-	if err != nil {
-		return err
-	}
-	err = os.RemoveAll(f.FilePathJoin(f.UserHomeDir(), ".flyte", "k3s"))
-	if err != nil {
-		return err
-	}
-	return nil
-}
 
 // GetSandbox will return sandbox container if it exist
 func GetSandbox(ctx context.Context, cli Docker) *types.Container {
@@ -141,12 +111,12 @@ func StartContainer(ctx context.Context, cli Docker, volumes []mount.Mount, expo
 
 // WatchError will return channel for watching errors of a container
 func WatchError(ctx context.Context, cli Docker, id string) (<-chan container.ContainerWaitOKBody, <-chan error) {
-	return cli.ContainerWait(context.Background(), id, container.WaitConditionNotRunning)
+	return cli.ContainerWait(ctx, id, container.WaitConditionNotRunning)
 }
 
 // ReadLogs will return io scanner for reading the logs of a container
 func ReadLogs(ctx context.Context, cli Docker, id string) (*bufio.Scanner, error) {
-	reader, err := cli.ContainerLogs(context.Background(), id, types.ContainerLogsOptions{
+	reader, err := cli.ContainerLogs(ctx, id, types.ContainerLogsOptions{
 		ShowStderr: true,
 		ShowStdout: true,
 		Timestamps: true,
@@ -167,7 +137,7 @@ func WaitForSandbox(reader *bufio.Scanner, message string) bool {
 			fmt.Printf("Register all flytesnacks example by running 'flytectl register examples  -d development  -p flytesnacks' \n")
 			fmt.Printf("Add KUBECONFIG and FLYTECTL_CONFIG to your environment variable \n")
 			fmt.Printf("export KUBECONFIG=%v \n", Kubeconfig)
-			fmt.Printf("export FLYTECTL_CONFIG=%v \n", FlytectlConfig)
+			fmt.Printf("export FLYTECTL_CONFIG=%v \n", util.FlytectlConfig)
 			return true
 		}
 		fmt.Println(reader.Text())
